package/delve: only select host-go when needed
Commit 58a291470acf (package/pkg-golang: select BR2_PACKAGE_HOST_GO)
ensured that we had the host-go enabled for all packages that need it.

However, for delve, the select was added to the _ARCH_SUPPORTS symbol,
which means that host-go is built as soon as delve supports the target,
even when devle itself is not enabled.

Move the select to the main symbol.

Fixes: 58a291470acf458239202bd81bfc642b9cadd298

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>