configs: stm32f429_disco_xip: use same headers as kernel
Instead of downloading a different version of the kernel for the headers as
for the runtime kernel itself, use the same one.  By doing so, we now need
to add the hash file for linux-headers.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit a4d38ef61cb1846871d15f49bdffd70a5ca05a2c)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>