board/xilinx: re-add hash file for xilinx-prebuilt
When xilinx pre-built was added in a6648ad8da (boot/xilinx-prebuilt:
new boot package), the hash file was initially a board-specific hash
file. However, that hash file, in commit a6648ad8da, really belonged
to the xilinx-prebuilt package, for two reasons: the package needs a
hash file for the version it installs, and no defconfig was using it
at that time. So the hash file was moved out of the boar directory,
and into the package directory when the patch was applied.

When the xilinx-related boards were changed to use xilinx-prebuilt,  in
551ba16a0f (configs/zynqmp_*: migrate to xilinx-prebuilt) and 8271323649
(configs/versal_vck190_defconfig: migrate to xilinx-prebuilt), they were
correctly made to use a custom, non-default version. But then, it meant
that those defconfig would no longer have a hash when we would bump the
default version in xilinx-prebuilt.

Re-add the initial board-specific hash file now that it is required.

Fixes: 551ba16a0f09fe550acb3ef6d12a1b2587d03292
Fixes: 8271323649c42420adaf4513b0472a7203f7784d

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Neal Frager <neal.frager@amd.com>
Reviewed-by: Neal Frager <neal.frager@amd.com>