package/go-bootstrap-stage3: move missing patch
The patch was added in commit 25991c843c3f (package/go-bootstrap-stage3:
fix go-bootstrap when parent dir contains invalid .git). Then all the
go-compiler packages were moved to a common sub-directory, in commit
0290c543dec8 (package/go: new subdirectory for go variants).

However, the patch for commit 0290c543dec8 was sent to the mailing list
before 25991c843c3f was applied, so it could not have moved a file that
was only added later.

Move that file, now.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Thomas Perale <thomas.perale@mind.be>
Cc: Romain Naour <romain.naour@smile.fr>