Revert "package/xz: convert to cmake build"
The conversion to cmake was supposed to be an RFC, and was not supposed
to be applied [0], as the cmake support is not yet complete, while the
autotools buildsystem is still functional [1].

The host-xz uses the target options, which is plain wrong. The target
options forcubly disable shared libraries for no apparent reason, which
causes the host build to also disable shared libs, which in turn causes
host-python3 to fails to build. host-python3 has no dependency on
host-xz, but it autodetects its presence, so if the build order caused
host-xz to be built before host-python3, its presence will be detected
(this needs to be fixed in host-python3 anyway, though) [2].

It also creates build issues in our reference envirnoment [3]

Finally, there was a typo in the CXX override:
    -DCMAKE_CXX_COMPILER_LAUNCHER==""

This reverts commit 3f464c2e2826f02d9ee6f738df67c078e37cb327.

Fixes: #5

[0] https://lore.kernel.org/buildroot/20240626103623.7d7412c0@gmx.net/
[1] https://tukaani.org/xz/#_building_from_xz_git
[2] https://gitlab.com/buildroot.org/buildroot/-/issues/5
[3] https://lore.kernel.org/buildroot/ZnqUPBAcDbB7ND64@tl-lnx-nyma7486-2/

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>