package/sox: fix patch fuzz
Commit 8f88a644ed "support/scripts/apply-patches.sh: set the maximum
fuzz factor to 0" reduced the fuzz factor.

Due to this change, sox fails to build with output:

    Applying 0001-uclibc.patch using patch:
    patching file src/formats.c
    Hunk #1 FAILED at 409.
    Hunk #2 succeeded at 477 (offset 60 lines).
    1 out of 2 hunks FAILED -- saving rejects to file src/formats.c.rej

This commit refreshes the package patch on the current package version.
Note 1: the failing patch 0001 was regenerated with "git format-patch".
Since the patch file 0001 name changed in the regeneration, the
".checkpackageignore" entry is also updated accordingly.
Note 2: patch 0002 did not require any update.

Fixes:
http://autobuild.buildroot.org/results/1bc396386aa3321fe48101caf9635ae41feb057b

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>