package/nmap: bump to version 7.95
For change log, see:
https://nmap.org/changelog.html#7.95

A notable change is the switch from pcre to pcre2. This change is
reflected in Config.in and nmap.mk.

This commit also adds the sha256 hash, published at:
https://nmap.org/dist/sigs/nmap-7.95.tar.bz2.digest.txt

The package patches are also rebased on this new version. The nmap
package was failing to build since commit 8f88a644ed
"support/scripts/apply-patches.sh: set the maximum fuzz factor to 0".
The patch 0001 was regenerated with "git format-patch". This operation
renamed the patch file. This change is reflected in the
".checkpackageignore" entry.

Fixes:
http://autobuild.buildroot.org/results/417ef0b2a80e8ec5859813a9f16f773881899025

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>