package/rt-tests: fix patch fuzz
Commit 8f88a644ed "support/scripts/apply-patches.sh: set the maximum
fuzz factor to 0" reduced the fuzz factor.

Due to this change, rt-tests fails to build with output:

    Applying 0001-Fix-a-build-issue-with-uClibc-ng.patch using patch:
    patching file src/cyclictest/cyclictest.c
    Hunk #1 FAILED at 58.
    1 out of 1 hunk FAILED -- saving rejects to file src/cyclictest/cyclictest.c.rej

This commit refreshes the package patches on the current package version.

The patch 0002 had an "Upstream:" tag in brackets. The tag was not
detected as such by check-package. This commit removes those brackets
and also removes the ".checkpackageignore" entry.

Fixes:
http://autobuild.buildroot.org/results/3e0a3b8d8c42d8e82ecf47b80bca875de739af5b

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>