package/xfsprogs: fix patch fuzz
Commit 8f88a644ed "support/scripts/apply-patches.sh: set the maximum
fuzz factor to 0" reduced the fuzz factor.

Due to this change, xfsprogs fails to build with output:

    Applying 0001-mdrestore-do-not-do-dynamic-linking-of-libtool-libra.patch using patch:
    patching file mdrestore/Makefile
    Hunk #1 FAILED at 10.
    1 out of 1 hunk FAILED -- saving rejects to file mdrestore/Makefile.rej
    patching file scrub/Makefile
    Hunk #1 FAILED at 70.
    1 out of 1 hunk FAILED -- saving rejects to file scrub/Makefile.rej

This commit refreshes the package patches on the current package
version. Note: the patch 0001 file name is changed because it was
regenerated with the "git format-patch" command.

The patch 0001 had an "Upstream:" tag in brackets. The tag was not
detected as such by check-package. This commit updates this tag. The
patch 0002 is also marked with an alternative upstream commit, so it
can be easily dropped at the next bump.

With those changes, the ".checkpackageignore" entries are no longer
needed, so this commit removes those.

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>