package/lxc: update the patches to be applied with fuzz 0
Commit 8f88a644ed7d6 ("support/scripts/apply-patches.sh: set the maximum
fuzz factor to 0") reduced the fuzz factor.

Due to this change, lxc fails to build with output:

    Applying 0001-src-lxc-syscall_numbers.h-drop-define-1.patch using patch:
    patching file src/lxc/syscall_numbers.h
    Hunk #1 FAILED at 53.
    Hunk #2 FAILED at 92.
    Hunk #3 FAILED at 129.
    Hunk #4 FAILED at 166.
    Hunk #5 FAILED at 203.
    Hunk #6 FAILED at 272.
    Hunk #7 FAILED at 309.
    Hunk #8 FAILED at 346.
    Hunk #9 FAILED at 383.
    Hunk #10 FAILED at 440.
    Hunk #11 FAILED at 477.
    Hunk #12 FAILED at 518.
    Hunk #13 FAILED at 759.
    13 out of 13 hunks FAILED -- saving rejects to file src/lxc/syscall_numbers.h.rej

This commit refreshes the package patch on the current package version.

As you can see from [1], the patch was created for version v6.0.0 but
has been backported in Buildroot without being rebased on lxc v5.0.0,
where loongarch64 support was not yet merged. This is further evidence
of the importance of applying patches with fuzz 0.

[1] https://github.com/lxc/lxc/commit/82fe01821cd5cb8548598d7d93b07d6ef3f6b604
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Romain Naour <romain.naour@smile.fr>