package/rdesktop: update the patches to be applied with fuzz 0
Commit 8f88a644ed "support/scripts/apply-patches.sh: set the maximum
fuzz factor to 0" reduced the fuzz factor.

Due to this change, rdesktop fails to build with output:

    Applying 0001-8bit-colors.patch using patch:
    patching file xwin.c
    Hunk #1 succeeded at 1801 (offset 340 lines).
    Hunk #2 FAILED at 1568.
    1 out of 2 hunks FAILED -- saving rejects to file xwin.c.rej

This commit refreshes the package patches on the current package
version. The original patch was not generated with "git format-patch",
and had no information (no commit log, no author, no date, ...). Since
it was introduced in commit [1], the author and date is set to this
commit. The "Upstream:" tag is also added to the patch, mentioning
the upstream project is reported as unmaintained on its homepage (at
the time of this commit). Finally, the ".checkpackageignore" entry is
removed, since it is no longer needed.

[1] https://gitlab.com/buildroot.org/buildroot/-/commit/e2dde9c7cdc2f9ed851acfb7daa858ffa1744246

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Romain Naour <romain.naour@smile.fr>