configs/visionfive2: fix build by adding kernel header version
commit d9e6d2d081 "configs/visionfive2: update kernel to upstream
6.8.2" removed the config directive:

    BR2_PACKAGE_HOST_LINUX_HEADERS_CUSTOM_5_15=y

but forgot to add:

    BR2_PACKAGE_HOST_LINUX_HEADERS_CUSTOM_6_8=y

It was most likely generated with a "make savedefconfig", while Linux
kernel v6.8 was the default at that time. After commit aa70c331a2
"linux: bump latest version to 6.9", the defconfig failed to build
with error:

    Incorrect selection of kernel headers: expected 6.9.x, got 6.8.x

This commit fixes the defconfig by adding back the kernel custom header
version.

Fixes:
https://gitlab.com/buildroot.org/buildroot/-/jobs/7271150175

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>