package/gpsd: move python option later
It's in the middle of options that are specific to the daemon, while
it itself is not specific to the daemon. We're going to add an option
to enable/disable the daemon soon, so that python option would get
hidden away.

Move it further down, out of scope of the daemon-related options.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>
Cc: Bernd Kuhls <bernd@kuhls.net>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>