package/pppd: add patch fixing build issue on SPARC/SPARC64
Integrate a patch, which was submitted upstream, to work-around an
issue with the SPARC kernel headers. The kernel headers have been
fixed in the upstream kernel (in the to-be-released 6.10 kernel), but
we need a workaround for the older toolchains, until we believe 6.10
is "old" enough to be able to drop the workaround.

Fixes:

  http://autobuild.buildroot.net/results/272c464ed4f9392535fa3b7613218dbd03acf901/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit b162aab7c6da6b45dc5ba9ad30b59b4362e2dc2e)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>