support/testing/infra/emulator.py: add license header
Since we copied a piece from code from pexpect, we need to also follow
the license of pexpect. Effectively, that makes the file licensed under
both our default GPL-2.0 and pexpect's ISC license. This is already
hinted at in the comment inside _repl_sh_child, but let's make it
explicit by adding a license header to the file.

Ideally, of course, each file in Buildroot would have a license file.

This was done with the reuse tool:
    reuse annotate -l GPL-2.0 -l ISC support/testing/infra/emulator.py

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>